Moved utils to src module and imported it to app.py
